refactor: extract error handler to useErrorHandler
